Move metrics to Shared
